fix: refactor VersionList
