+++ src/components/Nodes/ClientConfig/VersionList/VersionListItem.js
@@ -7,8 +7,8 @@ import Typography from '@material-ui/core/Typography'
 import CloudDownloadIcon from '@material-ui/icons/CloudDownload'
 import CheckBoxIcon from '@material-ui/icons/CheckBox'
 import styled, { css } from 'styled-components'
-import Spinner from '../../shared/Spinner'
-import { without } from '../../../lib/utils'
+import Spinner from '../../../shared/Spinner'
+import { without } from '../../../../lib/utils'
 
 export default class VersionListItem extends Component {
   static propTypes = {
@@ -126,9 +126,7 @@ export default class VersionListItem extends Component {
     return (
       <StyledListItem
         button
-        onClick={() => {
-          this.handleReleaseSelect(release)
-        }}
+        onClick={() => this.handleReleaseSelect(release)}
         selected={isSelectedRelease(release)}
         isDownloading={isDownloading}
         alt={release.name}
